backend/requirements.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.6.0
Flask==1.1.1
idna==2.9
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
openapi-spec-validator==0.2.8
Pillow==7.0.0
pyrsistent==0.15.7
PyYAML==5.3
requests==2.23.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
